Replace CancelSubscriptionCommand with SubscriberService
